backend/requirements.txt
fastapi
uvicorn[standard]
firebase-admin
sqlalchemy
databases[mysql]
aiomysql
pymysql
pydantic
python-dotenv
